Add a description for h2c
I came across this addition to the spec in both nutshell, and in cashu-ts (v1). Maybe this is not ready to get added to the spec, but I figured I would submit a PR anyway for fun.

Also, when I first implemented the BDHKE flow a couple of months ago, the hash_to_curve function took me a while to figure out what it was. Maybe this is because I'm a noob, but I figured a clear explanation would be helpful for future nutty noobs.