Add a description for h2c (#88)
* Add a description for h2c

I came across this addition to the spec in both nutshell, and in cashu-ts (v1). Maybe this is not ready to get added to the spec, but I figured I would submit a PR anyway for fun.

Also, when I first implemented the BDHKE flow a couple of months ago, the hash_to_curve function took me a while to figure out what it was. Maybe this is because I'm a noob, but I figured a clear explanation would be helpful for future nutty noobs.

* fix h2c description

Before I was using the hash as the private key, then computing the public key. I think this is right now.

* Update 00.md

Co-authored-by: Angus Pearson <angus@toaster.cc>

---------

Co-authored-by: callebtc <93376500+callebtc@users.noreply.github.com>
Co-authored-by: Angus Pearson <angus@toaster.cc>